Use float.
git-svn-id: http://usg.lofar.org/svn/code/trunk@11444 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/PyCRTools/modules/tasks/crosscorrelateantennas.py
@@ -115,7 +115,7 @@ class FindPulseDelay(Task):
         maxpos=dict(default = lambda self: cr.hArray(int, self.trace.shape()[0]),
             doc = "position of pulse maximum.",
             output = True),
-        delay=dict(default = lambda self: cr.hArray(int, self.trace.shape()[0]),
+        delay=dict(default = lambda self: cr.hArray(float, self.trace.shape()[0]),
             doc = "pulse delays in seconds.",
             output = True),
     )
